Add files via upload
Bridging Data Gaps in Africa
Despite the global abundance of data, Africa continues to face critical data gaps—missing, fragmented, or inaccurate datasets that hinder effective policymaking and development. In today’s data-driven world, the lack of credible, timely, and relevant data results in misdirected efforts, inefficient resource allocation, and the exclusion of vulnerable populations. This project explores these challenges and underscores the need for reliable data to support inclusive, evidence-based governance. As highlighted by Paris21 and the Mo Ibrahim Foundation, data is central to understanding where we are, where we’re going, and how to get there efficiently.